feat(flights): refactor FlightsPage and FlightSearchPage to use new client components
